Added AppModule for hilt dependency injection.
