+++ app/build.gradle
@@ -121,6 +121,6 @@ dependencies {
     ksp "androidx.room:room-compiler:2.6.1"
 
     // Hilt
-    implementation "com.google.dagger:hilt-android:2.44"
-    kapt "com.google.dagger:hilt-compiler:2.44"
+    implementation "com.google.dagger:hilt-android:2.48"
+    kapt "com.google.dagger:hilt-compiler:2.48"
 }
