+++ app/src/main/java/com/developerspace/webrtcsample/util/Constants.kt
@@ -1,0 +1,5 @@
+package com.developerspace.webrtcsample.util
+
+object Constants {
+    const val APP_DATABASE_NAME = "app_database"
+}
