+++ app/src/main/java/com/developerspace/webrtcsample/util/db/AppDatabase.kt
@@ -1,37 +1,15 @@
 package com.developerspace.webrtcsample.util.db
 
-import android.content.Context
 import androidx.room.Database
-import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.developerspace.webrtcsample.compose.ui.viewmodel.RecentMessage
+import com.developerspace.webrtcsample.util.db.dao.UserDao
+import com.developerspace.webrtcsample.util.db.entity.UserData
 
 @Database(
     entities = [UserData::class],
     version = 1
 )
 abstract class AppDatabase: RoomDatabase() {
-
-    abstract val userDao: UserDao
-
-    companion object {
-        // Singleton prevents multiple instances of database opening at the
-        // same time.
-        @Volatile
-        private var INSTANCE: AppDatabase? = null
-
-        fun getDatabase(context: Context): AppDatabase {
-            // if the INSTANCE is not null, then return it,
-            // if it is, then create the database
-            return INSTANCE ?: synchronized(this) {
-                val instance = Room.databaseBuilder(
-                    context.applicationContext,
-                    AppDatabase::class.java,
-                    "app_database"
-                ).build()
-                INSTANCE = instance
-                // return instance
-                instance
-            }
-        }
-    }
+    abstract fun userDao(): UserDao
 }
