+++ app/src/main/java/com/developerspace/webrtcsample/util/db/entity/UserData.kt
@@ -1,0 +1,13 @@
+package com.developerspace.webrtcsample.util.db.entity
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "user_data")
+data class UserData(
+    @PrimaryKey @ColumnInfo(name = "user_id") val userId: String = "",
+    @ColumnInfo(name = "user_name") val userName: String? = "",
+    @ColumnInfo(name = "profile_url") val profileUrl: String? = "",
+    @ColumnInfo(name = "online_status") val onlineStatus: Boolean? = false
+)
