+++ build.gradle
@@ -20,7 +20,7 @@ buildscript {
 
 plugins {
     id 'com.google.devtools.ksp' version '1.9.21-1.0.16' apply false
-    id 'com.google.dagger.hilt.android' version '2.44' apply false
+    id 'com.google.dagger.hilt.android' version '2.48' apply false
 }
 
 allprojects {
